Replace long with int
Turns out that we store these values as int(11) in MySql.
The 11 is irrelevant and pertains to display. int is a 4 byte
(aka 32 bit) integer. So this maps to a .NET Int32 (aka int).

While changing keeping it long might be future proofing, it also
requires changes to GHfW and I figure let's jump that hurdle
when we get there.

+++ Octokit.Reactive/IObservableSshKeysClient.cs
@@ -13,7 +13,7 @@ namespace Octokit.Reactive
         /// <param name="id">The ID of the SSH key.</param>
         /// <returns>A <see cref="SshKey"/></returns>
         [SuppressMessage("Microsoft.Naming", "CA1716:IdentifiersShouldNotMatchKeywords", MessageId = "Get")]
-        IObservable<SshKey> Get(long id);
+        IObservable<SshKey> Get(int id);
 
         /// <summary>
         /// Retrieves the <see cref="SshKey"/> for the specified id.
@@ -46,7 +46,7 @@ namespace Octokit.Reactive
         /// <param name="key"></param>
         /// <exception cref="AuthenticationException">Thrown if the client is not authenticated.</exception>
         /// <returns>A <see cref="User"/></returns>
-        IObservable<SshKey> Update(long id, SshKeyUpdate key);
+        IObservable<SshKey> Update(int id, SshKeyUpdate key);
 
         /// <summary>
         /// Update the specified <see cref="UserUpdate"/>.
@@ -54,6 +54,6 @@ namespace Octokit.Reactive
         /// <param name="id">The id of the SSH key</param>
         /// <exception cref="AuthenticationException">Thrown if the client is not authenticated.</exception>
         /// <returns>A <see cref="User"/></returns>
-        IObservable<Unit> Delete(long id);
+        IObservable<Unit> Delete(int id);
     }
 }
